+++ src/botPage/view/deriv/components/blockly-workspace/index.js
@@ -1,0 +1,3 @@
+import BlockLyWorkspace from "./blockly-workspace.jsx";
+
+export default BlockLyWorkspace;
