+++ src/botPage/view/deriv/layout/Main/main.jsx
@@ -11,9 +11,9 @@ import Footer from "../Footer";
 import Header from "../Header";
 import ToolBox from "../ToolBox";
 import SidebarToggle from '../../components/SidebarToggle';
+import BlockLyWorkspace from '../../components/blockly-workspace';
 
-
-const Main = ({blockly}) => {
+const Main = ({ blockly }) => {
   const { should_show_tour } = useSelector((state) => state.ui);
   const dispatch = useDispatch();
   React.useEffect(() => {
@@ -24,17 +24,18 @@ const Main = ({blockly}) => {
   }, []);
   return (
     <div className="main">
+      <BlockLyWorkspace blockly={blockly}/>
       {should_show_tour && <Tour />}
       <Header />
       <ToolBox blockly={blockly} />
       <SidebarToggle />
-      <Footer/>
+      <Footer />
     </div>
   );
 };
 
 Main.propTypes = {
-    blockly: PropTypes.object.isRequired,
-  }
+  blockly: PropTypes.object.isRequired,
+}
 export default Main;
